requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
pyyaml>=6.0.0
requests>=2.31.0
pdfplumber>=0.10.0
tabulate>=0.9.0
colorlog>=6.7.0
